commit may 5, 2017 14:10

+++ templates/default/app/loaddetailretailload.tpl.php
@@ -541,7 +541,11 @@ pre(array('$vars'=>$vars));
 
 			if(name=='retail_processingfee') {
 				var retail_processingfee = myForm.getItemValue('retail_processingfee');
-				console.log({onBlur:retail_processingfee,type:typeof(retail_processingfee),value:parseInt(retail_processingfee)});
+				console.log({onBlur:retail_processingfee,type:typeof(retail_processingfee),value:parseFloat(retail_processingfee)});
+
+				//if(parseInt(retail_processingfee)) {
+
+				//}
 			}
 
 		});
